New files:  index and get-sakai
The index file serves as a table of contents of all files in this
directory, which is the root directory in this case.  Eventually, we
will have sub-directories and I recommend that we have an index file for
each of them.

I have also taken a crack at a "Get Sakai" file.  Currently, it links to
the installation instructions on Confluence for 2.9.  We should discuss
if release documentation will moved to the cle-documentation repo in the
future.