Guidelines Document
I have created a set of guidelines that I hope will make it easier to
contribute to this documentation effort in a manner that provides some
structure and enables the later development of various deployment
strategies.

In particular, this document calls for three important elements:

1.  Documents shall be in strict XHTML.  This allows docs to be parsed
easily.
2.  An author DIV identifies the primary authors and subsequent
contributors.
3.  A version DIV identifies the first version of Sakai this applies to.

In addition to the above, I have made a statement about copyright that
assigns all rights to this material to the Apereo Foundation and
requires that all contributors sign an Apereo CLA first.